components/Sensor/requirements.txt
csrbuilder==0.10.1
oscrypto==1.1.0
asn1crypto==1.2.0
pycrypto==2.6.1
pycryptodomex==3.9.4
pyOpenSSL==19.1.0
certvalidator==0.11.1

